split up tests into seperate files
